Ran migrations

+++ spec/dummy/db/schema.rb
@@ -2,8 +2,8 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# This file is the source Rails uses to define your schema when running `rails
-# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
 # be faster and is potentially less error prone than running all of your
 # migrations from scratch. Old migrations may fail to apply correctly if those
 # migrations use external dependencies or application code.
@@ -637,8 +637,8 @@ ActiveRecord::Schema.define(version: 2020_12_24_123911) do
   end
 
   create_table "stripe_subscription_default_tax_rates", force: :cascade do |t|
-    t.integer "stripe_subscription_id", null: false
-    t.integer "stripe_tax_rate_id", null: false
+    t.bigint "stripe_subscription_id", null: false
+    t.bigint "stripe_tax_rate_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["stripe_subscription_id"], name: "index_on_subscription"
